2005-07-20  Zoltan Varga  <vargaz@freemail.hu>
	* mini-sparc.c (mono_arch_get_vcall_slot_addr): Fix warning.

	* tramp-sparc.c (mono_arch_create_trampoline_code): Fix order of
	trampoline parameters.

svn path=/trunk/mono/; revision=47446


Commit migrated from https://github.com/mono/mono/commit/a2e49f320b05cc327ae1a2364f92cd69cea9c917

+++ src/mono/mono/mini/mini-sparc.c
@@ -2171,7 +2171,7 @@ mono_sparc_is_virtual_call (guint32 *code)
 gpointer*
 mono_arch_get_vcall_slot_addr (guint8 *code8, gpointer *regs)
 {
-	guint32 *code = (guint8*)(gpointer)code8;
+	guint32 *code = (guint32*)(gpointer)code8;
 	guint32 ins = code [0];
 	guint32 prev_ins = code [-1];
